Нормализована работа избранного
А так же переименован слайс offer-in-detail, добавлен универсальный компонент для добавления в избранное, убраны остатки тега <a>
