+++ src/components/footer/footer.tsx
@@ -1,10 +1,13 @@
+import { Link } from 'react-router-dom';
+import { AppRoute } from '../../constants';
+
 function Footer() {
   return (
 
     <footer className="footer container">
-      <a className="footer__logo-link" href="main.html">
+      <Link className="footer__logo-link" to={AppRoute.Main}>
         <img className="footer__logo" src="img/logo.svg" alt="6 cities logo" width="64" height="33" />
-      </a>
+      </Link>
     </footer>
   );
 }
