+++ src/components/offer-loading/offer-loading.tsx
@@ -1,4 +1,6 @@
+import { Link } from 'react-router-dom';
 import './loader.css';
+import { AppRoute } from '../../constants';
 
 export default function OfferLoading(): JSX.Element {
 
@@ -9,18 +11,18 @@ export default function OfferLoading(): JSX.Element {
           <div className="container">
             <div className="header__wrapper">
               <div className="header__left">
-                <a className="header__logo-link header__logo-link--active" href="/:city">
+                <Link className="header__logo-link header__logo-link--active" to={AppRoute.Main}>
                   <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
-                </a>
+                </Link>
               </div>
               <nav className="header__nav">
                 <ul className="header__nav-list">
                   <li className="header__nav-item user">
-                    <a className="header__nav-link header__nav-link--profile" href="/login">
+                    <Link className="header__nav-link header__nav-link--profile" to={AppRoute.Login}>
                       <div className="header__avatar-wrapper user__avatar-wrapper">
                       </div>
                       <span className="header__login">Sign in</span>
-                    </a>
+                    </Link>
                   </li>
                 </ul>
               </nav>
