--- src/components/settings/settings.tsx
@@ -1,11 +1,0 @@
-export const data = {
-  cities: [
-    { id: 'paris', name: 'Paris' },
-    { id: 'cologne', name: 'Cologne' },
-    { id: 'brussels', name: 'Brussels' },
-    { id: 'amsterdam', name: 'Amsterdam' },
-    { id: 'hamburg', name: 'Hamburg' },
-    { id: 'dusseldorf', name: 'Dusseldorf' }]
-  ,
-  hasNavigation: true
-};
