+++ src/components/sorting-block/sorting-block.tsx
@@ -2,7 +2,7 @@ import { memo, useState } from 'react';
 import { SortingArray } from '../../constants';
 import { SortingType } from '../../types';
 import { useAppDispatch } from '../../hooks';
-import { setSortingMode } from '../../store/slices/offer-slice';
+import { setSortingMode } from '../../store/slices/offers-slice';
 
 function SortingBlock() {
   const [isBlockOpened, setBlockState] = useState(false);
