+++ src/constants.ts
@@ -66,3 +66,15 @@ export const SortingArray: SortingType[] = [
     value: 'Top rated first'
   }
 ];
+
+export const citiesData = {
+  cities: [
+    { id: 'paris', name: 'Paris' },
+    { id: 'cologne', name: 'Cologne' },
+    { id: 'brussels', name: 'Brussels' },
+    { id: 'amsterdam', name: 'Amsterdam' },
+    { id: 'hamburg', name: 'Hamburg' },
+    { id: 'dusseldorf', name: 'Dusseldorf' }]
+  ,
+  hasNavigation: true
+};
